Bug#37201922 Tables with 256 disk columns will not use disk scan

+++ storage/ndb/src/ndbapi/NdbDictionaryImpl.hpp
@@ -315,7 +315,7 @@ class NdbTableImpl : public NdbDictionary::Table, public NdbDictObjectImpl {
   // if all pk = dk then this is zero!
   Uint8 m_noOfDistributionKeys;
   Uint8 m_noOfBlobs;
-  Uint8 m_noOfDiskColumns;
+  Uint16 m_noOfDiskColumns;
   Uint8 m_replicaCount;
 
   /**
